+++ tests/internal/hyper/variables/groups/se3_tests.cpp
