+++ tests/internal/hyper/variables/groups/su2_tests.cpp
