+++ tests/internal/hyper/variables/intrinsics_tests.cpp
